Move check_* utils to own file
