+++ lib/streamlit/elements/vega_charts.py
@@ -1652,7 +1652,7 @@ class VegaChartsMixin:
             # Run some checks that are only relevant when selections are activated
 
             # Import here to avoid circular imports
-            from streamlit.elements.utils import (
+            from streamlit.elements.policies import (
                 check_cache_replay_rules,
                 check_callback_rules,
                 check_session_state_rules,
