+++ lib/streamlit/elements/widgets/select_slider.py
@@ -21,10 +21,12 @@ from typing import TYPE_CHECKING, Any, Callable, Generic, Sequence, Tuple, cast
 from typing_extensions import TypeGuard
 
 from streamlit.elements.form import current_form_id
-from streamlit.elements.utils import (
+from streamlit.elements.policies import (
     check_cache_replay_rules,
     check_callback_rules,
     check_session_state_rules,
+)
+from streamlit.elements.utils import (
     get_label_visibility_proto_value,
     maybe_coerce_enum,
     maybe_coerce_enum_sequence,
